# docs/swagger.yaml
...
info:
  contact: {}
  description: Сервис сбора метрик и алертинга
  title: Go Metrics
  version: "1.0"
paths:
...
  /update/counter/{name}/{value}:
    post:
      parameters:
      - description: Название метрики
        example: PollCount
        in: path
        name: name
        required: true
        type: string
      - description: Значение метрики
        example: "1"
        in: path
        name: value
        required: true
        type: string
      responses:
        "200":
          description: Метрику обновили
        "400":
          description: Некорректный запрос
        "404":
          description: Метрика не найдена
        "500":
          description: Ошибка
      summary: Обновление метрики типа Counter.
...
  /value/gauge/{name}:
    get:
      parameters:
      - description: Название метрики
        example: Sys
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Значение метрики
          schema:
            type: string
        "400":
          description: Неизвестный тип метрики
        "404":
          description: Метрика не найдена
        "500":
          description: Ошибка
      summary: Получить значение метрики типа Gauge.
swagger: "2.0"